spi: changed context managers on spice runner and manager separation
