+++ src/spiceInterface/spiceRunner.py
@@ -4,7 +4,8 @@ This is the only class allowed to call spice.
 This is the only class that knows Spice File Manager
 """
 
-from .spiceFileManager import SpiceFileManager
+from .spiceFileReader import SpiceFileReader
+from .spiceFileWriter import SpiceFileWriter
 from ..circuit.components import LET
 from ..simconfig.simulationConfig import sim_config
 from ..utils.math import InDir
@@ -13,15 +14,109 @@ from typing import Callable
 from abc import ABC
 
 
+class MonteCarlo:
+    """
+    Context Mangers that sets the number of MC simulations.
+    """
+
+    def __init__(self, file_manager: SpiceFileWriter, num_testes: int):
+        self.file_manager: SpiceFileWriter = file_manager
+        self.num = num_testes
+
+    def __enter__(self):
+        self.file_manager.set_monte_carlo(self.num)
+
+    def __exit__(self, type, value, traceback):
+        self.file_manager.set_monte_carlo(0)
+
+
+class SET:
+    """
+    Context Mangers that sets the number a single fault.
+    """
+
+    def __init__(self, file_manager: SpiceFileWriter, let: LET, current: float = None):
+        self.file_manager: SpiceFileWriter = file_manager
+        self.let = let
+        if current == None:
+            self.current = let.current
+        else:
+            self.current = current
+
+    def __enter__(self):
+        self.file_manager.set_pulse(self.let, self.current)
+        self.file_manager.measure_pulse(self.let.node_name, self.let.output_name)
+
+    def __exit__(self, type, value, traceback):
+        pass
+        # self.file_manager.set_pulse(self.let, 0)
+
+
+class Vss:
+    """
+    Context Manager that sets the vss of the simulation.
+    """
+
+    def __init__(self, file_manager: SpiceFileWriter, vss: float):
+        self.file_manager: SpiceFileWriter = file_manager
+        self.vss = vss
+
+    def __enter__(self):
+        self.file_manager.set_vss(self.vss)
+
+    def __exit__(self, type, value, traceback):
+        pass
+
+
+class Vdd:
+    """
+    Context Manager that sets the vdd of the simulation.
+    """
+
+    def __init__(self, file_manager: SpiceFileWriter, vdd: float):
+        self.file_manager: SpiceFileWriter = file_manager
+        self.vdd = vdd
+
+    def __enter__(self):
+        self.file_manager.set_vdd(self.vdd)
+
+    def __exit__(self, type, value, traceback):
+        pass
+
+
+class Inputs:
+    """
+    Context Mangers that sets the input signals of the simulation.
+    """
+
+    def __init__(
+        self, file_manager: SpiceFileWriter, inputs: list, vdd: float, vss: float = 0
+    ):
+        """
+        Constructor.
+
+            inputs (list[int]): Input logical values, either 0 or 1.
+            vdd (float): Input vdd.
+            vdd (float): Input vss.
+        """
+        self.file_manager: SpiceFileWriter = file_manager
+        self.vdd = vdd
+        self.vss = vss
+        self.inputs = {}
+        for entrada in inputs:
+            self.inputs[entrada.name] = entrada.signal
+
+    def __enter__(self):
+        self.file_manager.set_signals(self.inputs, self.vdd, self.vss)
+
+    def __exit__(self, type, value, traceback):
+        pass
+
+
 class SpiceRunner(ABC):
     """
     Responsible for running spice. Used as a Interface for spice to the rest of the system.
     """
-
-    # TODO This is a huge problem, this variable is a class variable not a object variable
-    # I want you to be able to call with SpiceRunner().Vdd(0.5) for example but for that the Vdd context manager
-    # Must know path_to_folder of the SpiceRunner instance, wich seems to be very hard to do
-    file_manager = None
 
     def __init__(
         self, spice_run_line: Callable, path_to_folder: str = "project"
@@ -35,97 +130,24 @@ class SpiceRunner(ABC):
         """
         self.path_to_folder = path_to_folder
         self.spice_run_line: Callable = spice_run_line
-        SpiceRunner.file_manager = SpiceFileManager(path_to_folder=path_to_folder)
-        self.file_manager = SpiceRunner.file_manager
+        self.file_manager = SpiceFileReader(path_to_folder=path_to_folder)
+        self.file_writer = SpiceFileWriter(path_to_folder=path_to_folder)
 
-    class Monte_Carlo:
-        """
-        Context Mangers that sets the number of MC simulations.
-        """
+    # Context Managers
+    def MonteCarlo(self, num_testes):
+        return MonteCarlo(self.file_writer, num_testes)
 
-        def __init__(self, num_testes):
-            self.num = num_testes
+    def SET(self, let: LET, current: float = None):
+        return SET(self.file_writer, let, current)
 
-        def __enter__(self):
-            SpiceRunner.file_manager.set_monte_carlo(self.num)
+    def Vss(self, vss: float):
+        return Vss(self.file_writer, vss)
 
-        def __exit__(self, type, value, traceback):
-            SpiceRunner.file_manager.set_monte_carlo(0)
+    def Vdd(self, vdd: float):
+        return Vdd(self.file_writer, vdd)
 
-    class SET:
-        """
-        Context Mangers that sets the number a single fault.
-        """
-
-        def __init__(self, let: LET, current: float = None):
-            self.let = let
-            if current == None:
-                self.current = let.current
-            else:
-                self.current = current
-
-        def __enter__(self):
-            SpiceRunner.file_manager.set_pulse(self.let, self.current)
-            SpiceRunner.file_manager.measure_pulse(
-                self.let.node_name, self.let.output_name
-            )
-
-        def __exit__(self, type, value, traceback):
-            pass
-            # SpiceRunner.file_manager.set_pulse(self.let, 0)
-
-    class Vss:
-        """
-        Context Manager that sets the vss of the simulation.
-        """
-
-        def __init__(self, vss: float):
-            self.vss = vss
-
-        def __enter__(self):
-            SpiceRunner.file_manager.set_vss(self.vss)
-
-        def __exit__(self, type, value, traceback):
-            pass
-
-    class Vdd:
-        """
-        Context Manager that sets the vdd of the simulation.
-        """
-
-        def __init__(self, vdd: float):
-            self.vdd = vdd
-
-        def __enter__(self):
-            SpiceRunner.file_manager.set_vdd(self.vdd)
-
-        def __exit__(self, type, value, traceback):
-            pass
-
-    class Inputs:
-        """
-        Context Mangers that sets the input signals of the simulation.
-        """
-
-        def __init__(self, inputs: list, vdd: float, vss: float = 0):
-            """
-            Constructor.
-
-                inputs (list[int]): Input logical values, either 0 or 1.
-                vdd (float): Input vdd.
-                vdd (float): Input vss.
-            """
-            self.vdd = vdd
-            self.vss = vss
-            self.inputs = {}
-            for entrada in inputs:
-                self.inputs[entrada.name] = entrada.signal
-
-        def __enter__(self):
-            SpiceRunner.file_manager.set_signals(self.inputs, self.vdd, self.vss)
-
-        def __exit__(self, type, value, traceback):
-            pass
+    def Inputs(self, inputs: list, vdd: float, vss: float = 0):
+        return Inputs(self.file_writer, inputs, vdd, vss)
 
     def _run_spice(self, labels: list = None) -> None:
         """
@@ -166,10 +188,10 @@ class SpiceRunner(ABC):
         Args:
             vdd (float): Standard vdd of the simulation.
         """
-        self.file_manager.set_vdd(vdd)
-        self.file_manager.set_vss(0)
-        self.file_manager.set_pulse(LET(0, vdd, "none", "none", [None, None]))
-        self.file_manager.set_monte_carlo(0)
+        self.file_writer.set_vdd(vdd)
+        self.file_writer.set_vss(0)
+        self.file_writer.set_pulse(LET(0, vdd, "none", "none", [None, None]))
+        self.file_writer.set_monte_carlo(0)
 
     def get_nodes(
         self, circ_name: str, tension_sources: list = None, inputs: list = None
@@ -185,7 +207,7 @@ class SpiceRunner(ABC):
         Returns:
             set: The label of all nodes.
         """
-        return SpiceRunner.file_manager.get_nodes(circ_name, tension_sources, inputs)
+        return self.file_manager.get_nodes(circ_name, tension_sources, inputs)
 
     def run_delay(self, input_name: str, output_name: str, inputs: list) -> float:
         """
@@ -201,14 +223,14 @@ class SpiceRunner(ABC):
         """
 
         # Set the signals to be simualted
-        SpiceRunner.file_manager.measure_delay(input_name, output_name, sim_config.vdd)
-        SpiceRunner.file_manager.set_signals(
+        self.file_writer.measure_delay(input_name, output_name, sim_config.vdd)
+        self.file_writer.set_signals(
             {input.name: input.signal for input in inputs}, sim_config.vdd
         )
         # Runs the simulation in the respective folder
         self._run_spice(["atraso_rr", "atraso_rf", "atraso_fr", "atraso_ff"])
         # Gets and returns the results
-        delay = SpiceRunner.file_manager.get_delay()
+        delay = self.file_manager.get_delay()
         return delay
 
     def run_SET(self, let: LET, current: float = None) -> tuple:
@@ -227,8 +249,8 @@ class SpiceRunner(ABC):
             # Runs the simulation
             self._run_spice(["minout", "maxout", "minnod", "maxnod"])
             # Gets the peak tensions in the node and output
-            peak_node = SpiceRunner.file_manager.get_peak_tension(let.orient[0], True)
-            peak_output = SpiceRunner.file_manager.get_peak_tension(let.orient[1])
+            peak_node = self.file_manager.get_peak_tension(let.orient[0], True)
+            peak_output = self.file_manager.get_peak_tension(let.orient[1])
         return (peak_node, peak_output)
 
     def run_pulse_width(self, let: LET, current: float = None) -> float:
@@ -243,9 +265,9 @@ class SpiceRunner(ABC):
             float: Fault width at output.
         """
         with self.SET(let, current):
-            SpiceRunner.file_manager.measure_pulse_width(let)
+            self.file_writer.measure_pulse_width(let)
             self._run_spice(["larg"])
-            output = SpiceRunner.file_manager.get_output()
+            output = self.file_manager.get_output()
 
         try:
             if output["larg"].value == None:
@@ -269,7 +291,7 @@ class SpiceRunner(ABC):
         measure_labels = [f"max{node}" for node in nodes] + [
             f"min{node}" for node in nodes
         ]
-        self.file_manager.measure_nodes(nodes)
+        self.file_writer.measure_nodes(nodes)
         self._run_spice(measure_labels)
         return self.file_manager.get_nodes_tension(nodes)
 
@@ -294,15 +316,13 @@ class SpiceRunner(ABC):
                 dist.var
             ] = f"gauss({dist.mean}, {dist.std_dev}, {dist.sigmas})"
 
-        with self.Monte_Carlo(sim_num):
+        with self.MonteCarlo(sim_num):
             self._run_spice()
 
         for dist, param in zip(distributions, params):
             sim_config.model_manager[dist.model][dist.var] = param
 
-        return SpiceRunner.file_manager.get_mc_instances(
-            sim_config.circuit.name, model_vars
-        )
+        return self.file_manager.get_mc_instances(sim_config.circuit.name, model_vars)
 
 
 class NGSpiceRunner(SpiceRunner):
@@ -343,7 +363,7 @@ if __name__ == "__main__":
     from ..circuit.circuit import Circuito
 
     sim_config.runner_type = NGSpiceRunner
-    TestManager = SpiceFileManager(path_to_folder=ptf)
+    TestManager = SpiceFileReader(path_to_folder=ptf)
     vdd = 0.9
     sim_config.runner.default(vdd)
